// src/enums/statuses.enum.ts
export enum Statuses {
  ONLINE = "ONLINE",
  OFFLINE = "OFFLINE",

  ACTIVE = "ACTIVE",
  UN_ACTIVE = "UN_ACTIVE",

  DELETED = "DELETED",
  UNDELETED = "UNDELETED",

  PENDING = "PENDING",
  CONFIRMED = "CONFIRMED",
  PREPARING = "PREPARING",
  SHIPPED = "SHIPPED",
  OUT_FOR_DELIVERY = "OUT_FOR_DELIVERY",
  DELIVERED = "DELIVERED",
  RETURNED = "RETURNED",
  CANCELED = "CANCELED",

  FAILED = "FAILED",
  REFUNDED = "REFUNDED",
  PAID = "PAID",

  INTERNAL = "INTERNAL",
  OUTSOURCED = "OUTSOURCED",

  DRAFT = "DRAFT",
  PENDING_REVIEW = "PENDING_REVIEW",
  REJECTED = "REJECTED",
  SUSPENDED = "SUSPENDED",
  CLOSED = "CLOSED",
}